Move transactionTypes loading to separate file
