+++ hedera-mirror-rest/server.js
@@ -40,7 +40,7 @@ const tokens = require('./tokens');
 const topicmessage = require('./topicmessage');
 const transactions = require('./transactions');
 const {handleError} = require('./middleware/httpErrorHandler');
-const {loadTransactionTypes} = require('./constants');
+const {loadTransactionTypes} = require('./transactionTypes');
 const {metricsHandler} = require('./middleware/metricsHandler');
 const {serveSwaggerDocs} = require('./middleware/openapiHandler');
 const {responseHandler} = require('./middleware/responseHandler');
